requirements/base.txt
# require lib

psycopg2-binary
click
peewee


# test, check lib

pytest
coverage